requirements.txt
pytest>=7.4.0
pytest-cov>=4.1.0
PyQt5>=5.15.9
openai>=1.3.5
python-dotenv>=1.0.0